requirements.txt
alembic==0.9.10
asn1crypto==0.24.0
aws==0.2.5
awscli==1.15.71
awsebcli==3.14.3
Babel==2.6.0
bcrypt==3.1.4
blessed==1.15.0
blinker==1.4
boto==2.49.0
botocore==1.10.70
cached-property==1.4.3
cement==2.8.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==2.3
docker==3.4.1
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
dominate==2.3.1
fabric==2.2.2
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.6
invoke==1.1.1
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
paramiko==2.4.1
pathspec==0.5.5
prettytable==0.7.2
pyasn1==0.4.4
pycparser==2.18
PyJWT==1.6.4
PyNaCl==1.2.1
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
SQLAlchemy==1.2.9
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
urllib3==1.22
visitor==0.1.3
wcwidth==0.1.7
websocket-client==0.48.0
Werkzeug==0.14.1
WTForms==2.2.1
